requirements.txt
fastapi==0.121.0
finnhub-python==2.4.25
httpx==0.28.1
pandas==2.3.3
plaid-python==37.1.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
redis==7.0.1
SQLAlchemy==2.0.44
starlette==0.49.3
uvicorn==0.38.0
yahooquery==2.4.1
stripe==14.1.0
psycopg2-binary==2.9.11
python-jose==3.5.0
pgvector==0.4.2
sec-parser==0.58.1
langgraph==1.0.6